codes/requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0

# Video processing
opencv-python>=4.8.0

# CLIP (for SAFR frame selection)
transformers>=4.45.0

# OneThinker / Qwen2.5-VL inference
vllm>=0.6.3
qwen-vl-utils>=0.0.8
accelerate>=0.26.0

# Utilities
tqdm>=4.65.0
scipy>=1.11.0
